[scudo] Add thread-safety annotations on TSD data members
Ideally, we want to assert that all the operations on
Cache/QuarantineCache always have the `Mutex` acquired. However,
the current architecture of accessing TSD is not easy to cooperate
with the thread-safety analysis because of pointer aliasing. In
alternative, we add the getters for accessing TSD member and attach
proper thread-safety annotations on them.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D142151

GitOrigin-RevId: ae1bd3adf02f4f381b3298b794fab8387d794538
Change-Id: Id8dff73beb3c87faa76f743891b024f09f9d95b0