[scudo] Mitigate page releasing thrashing
We have the heuristic to determine the threshold of doing page
releasing for smaller size classes. However, in a case that the
memory usage is bouncing between that threshold may result in
frequent try of page releasing but not returning much memory.

This CL add another heuristic to mitigate this problem by increasing
the minimum pages that potentially can be released. Note that this
heuristic is only applied on SizeClassAllocator64. SizeClassAllocator32
has a smaller group size so the overhead is smaller than 64-bit
platform.

Differential Revision: https://reviews.llvm.org/D144768

GitOrigin-RevId: 436ea5485d02c529e26a7a1007b82d581be016c4
Change-Id: Idcd9ad21ea2b41ab0e4e9a46d4b8e769bea31b6c